refactor to move xxhash64 into separate file
